📦 chore(sync-git-folders.sh): add script to sync git repositories and display progress bar
This commit adds a new script `sync-git-folders.sh` that can be used to sync multiple git repositories. The script takes a directory as an argument and finds all the git directories within it. It then pulls the latest changes and performs garbage collection for each repository in parallel, with a maximum number of parallel jobs specified by the `max_jobs` variable.

The script also includes a function `show_progress()` that displays a progress bar indicating the percentage of completion. The progress bar is updated for each repository being synced.

This script can be useful when you have multiple git repositories that need to be kept up to date and want to automate the process.